Fixed up tests after merge.
