+++ reducer/src/main/java/com/graphicsfuzz/reducer/reductionopportunities/IReductionOpportunityFinder.java
@@ -138,21 +138,6 @@ public interface IReductionOpportunityFinder<T extends IReductionOpportunity> {
     };
   }
 
-  static IReductionOpportunityFinder<CompoundToGuardReductionOpportunity> compoundToGuardFinder() {
-    return new IReductionOpportunityFinder<CompoundToGuardReductionOpportunity>() {
-      @Override
-      public List<CompoundToGuardReductionOpportunity> findOpportunities(ShaderJob shaderJob,
-                                                                         ReducerContext context) {
-        return CompoundToGuardReductionOpportunities.findOpportunities(shaderJob, context);
-      }
-
-      @Override
-      public String getName() {
-        return "compoundToGuard";
-      }
-    };
-  }
-
   static IReductionOpportunityFinder<SimplifyExprReductionOpportunity>
       inlineInitializerFinder() {
     return new IReductionOpportunityFinder<SimplifyExprReductionOpportunity>() {
